Implement MParticle.UploadAsync
Previously we had a partially asynchronous Upload method, which would force an immediate upload of all events in the the upload queue. This will split this method into a fully synchronous Upload method and a fully asynchronous UploadAsync method. Our former Upload implementation suffered from a bug in the testing code, which is why it ended up in a partially-synchronous state